refactor: 페이지 이동, 정리
