+++ pages/register/policy.tsx
@@ -1,9 +1,12 @@
 import Policy from 'client/pages/register/policy';
 
-export default function PolicyPage({}) {
+function PolicyPage() {
   return <Policy />;
 }
 
 export async function getServerSideProps() {
+  // TODO : server side 또는 static 적용
   return { props: {} };
 }
+
+export default PolicyPage;
